# backend/swagger.yaml
swagger: "2.0"
info:
  description: "APIs of Logistic Web Service"
  version: "1.0.0"
  title: "Logistic Web Service"
  contact:
    email: "d.bed11@gptgdt.net"
host: "localhost:8000"
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /admin/car:
    get:
      tags:
        - "car"
      summary: get cars
      produces:
        - application/json
      parameters:
        - in: query
          name: car_type
          type: string
          description: car_type of car (ex 8_ton, 20_ton)
        - in: query
          name: plate
          type: string
          description: plate of car 
        - in: query
          name: page
          type: string
          description: number of page 
        - in: query
          name: pageSize
          type: string
          description: number of page size 
      responses:
        200:
          description: get car successfully.
          schema:
            type: object
            example: {
                "plate": "string",
                "car_type": "string",
                "volumn": "string",
                "tonnage": "string",
            }
        400:
          description: car in not found.
      security:
        - bearerAuth: []
  /admin/car/{carId}:
    get:
      tags:
        - "car"
      summary: get car by id
      produces:
        - application/json
      parameters:
        - in: path
          name: carId
          type: string
          required: true
          description: Id of a car 
      responses:
        200:
          description: get car successfully.
          schema:
            type: object
            example: {
                "plate": "string",
                "car_type": "string",
                "volumn": "string",
                "tonnage": "string",
            }
        400:
          description: CarId is not found.
      security:
        - bearerAuth: []
    put:
      tags:
        - "car"
      summary: Admin update road 
      produces:
        - application/json
      consumes:
        - application/json
        - multipart/form-data
      parameters:
        - in: path
          name: carId
          type: string
          required: true
          description: Id of an car 
        - in: body
          name: road
          description: a new car.
          schema:
            type: object
            required:
              - volumn
              - tonnage 
              - car_type
            properties:
              volumn: 
                  type: number
              tonnage:
                  type: number
              car_type:
                type: string
                enum: ['8_ton', '20_ton']
        - in: formData
      responses:
        200:
          description: Update car successfully.
          schema:
            type: object
            example: {
                "plate": "string",
                "car_type": "string",
                "volumn": "string",
                "tonnage": "string",
            }
        400:
          description: road cannot be found.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "car"
      summary: "Delete a existing car"
      produces:
        - application/json
      parameters:
        - in: path
          name: carId 
          type: string
          required: true
          description: car Id
      responses:
        200:
          description: Delete car successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        400:
          description: Delete car failed.
      security:
        - bearerAuth: []
  /admin/car/create:
    post:
      tags:
        - "car"
      summary: "Create a new car"
      consumes:
        - application/json
      parameters:
        - in: body
          name: car
          description: a new car.
          schema:
            type: object
            properties:
              plate:
                type: string 
              car_type:
                type: string
                enum: ['8_ton', '20_ton']
              volumn:
                type: number 
              tonnage:
                type: number 
      responses:
        200:
          description: Create car successfully.
          schema:
            type: object
            example: {
                "plate": "string",
                "car_type": "string",
                "volumn": "string",
                "tonnage": "string",
            }
        400:
          description: Create car failed.
      security:
        - bearerAuth: []
...
  /applicant/{careerId}:
    post:
      tags:
        - "applicant"
      summary: "Create a new applicant"
      consumes:
        - multipart/form-data
        - application/json
      parameters:
        - in: path
          name: careerId
          type: string
          required: true
          description: Id of a career 
          required: true
        - in: formData
          name: firstName
          description: firstName of applicant.
          type: string 
          required: true
        - in: formData
          name: lastName
          description: lastName of applicant.
          required: true
        - in: formData
          name: phoneNumber
          description: phoneNumber of applicant.
          type: string 
          required: true
        - in: formData
          name: email
          description: email of applicant.
          type: string 
          required: true
        - in: formData
          name: source
          description: source of interest of applicant(staff, friend, email, phone, facebook, linkedin, search, event, other).
          type: string 
          required: true
        - in: formData
          name: message
          description: Custom message.
          type: string  
          required: true  
        - in: formData
          name: file
          description: CV file of applicant.
          type: file 
          required: true         
      responses:
        200:
          description: Create applicant successfully.
          schema:
            type: object
            example: {
              "firstName": "string",
              "lastName": "string",    
              "phoneNumber": "string",   
              "email" : "string",
              "source": "string",
              "message": "string",
              "status": "string",    
            }        
        400:
          description: Create applicant failed.
  /admin/applicant:
    get:
      tags:
        - "applicant"
      summary: get applicant
      produces:
        - application/json
      parameters:
        - in: query
          name: page
          type: string
          description: num of page
        - in: query
          name: pageSize
          type: string
          description: number of applicants in a page 
        - in: query
          name: sortBy
          type: string
          description: sort in order (firstName, lastName, -updatedAt)
        - in: query
          name: keyword
          type: string
          description: search by keyword
        - in: query
          name: department
          type: string
          description: filter by department
        - in: query
          name: type
          type: string
          description: filter by type
        - in: query
          name: location
          type: string
          description: filter by location
        - in: query
          name: state
          type: string
          description: filter by state
      responses:
        200:
          description: get applicant successfully.
          schema:
            type: object
            example: {
              "firstName": "string",
              "lastName": "string",    
              "phoneNumber": "string",   
              "email" : "string",
              "source": "string",
              "message": "string",
              "status": "string",       
            }          
        400:
          description: applicant is not found.  
      security:
        - bearerAuth: []
  /admin/applicant/{applicantId}:
    get:
      tags:
        - "applicant"
      summary: get applicant by id
      produces:
        - application/json
      parameters:
        - in: path
          name: applicantId
          type: string
          required: true
          description: Id of an applicant 
      responses:
        200:
          description: get applicant successfully.
          schema:
            type: object
            example: {
              "firstName": "string",
              "lastName": "string",    
              "phoneNumber": "string",   
              "email" : "string",
              "source": "string",
              "message": "string",
              "status": "string",       
            }          
        400:
          description: applicant is not found.
      security:
        - bearerAuth: []
    put:
      tags:
        - "applicant"
      summary: "Update a existing applicant status"
      consumes:
        - application/json
      parameters:
        - in: path
          name: applicantId
          schema: 
            type: string
          required: true
          description: applicant id
        - in: body
          name: applicant
          description: an existing applicant.
          schema:
            type: object      
            required: 
            - status   
            properties:
              status:
                type: string
                description: (approved, pending, rejected)                 
      responses:
        200:
          description: Update applicant successfully.
          schema:
            type: object
            example: {
              "firstName": "string",
              "lastName": "string",    
              "phoneNumber": "string",   
              "email" : "string",
              "source": "string",
              "message": "string",
              "status": "string",          
            }      
        400:
          description: Update applicant failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "applicant"
      summary: "Delete an existing applicant"
      produces:
        - application/json
      parameters:
        - in: path
          name: applicantId
          type: string
          required: true
          description: Id of a applicant 
      responses:
        200:
          description: Delete applicant successfully.
        400:
          description: Delete applicant failed.
      security:
        - bearerAuth: []
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"